refactor(url helper): layer url helper  separated
