+++ src/exporter/models/exporterStore.ts
@@ -9,7 +9,7 @@ import {
 import { Polygon } from '@turf/helpers';
 import { ApiHttpResponse } from '../../common/models/api-response';
 import { ResponseState } from '../../common/models/ResponseState';
-import EXPORTER_CONFIG from '../../common/config';
+import { getLayerUrl } from '../../common/helpers/layer-url';
 import { searchParams } from './search-params';
 import { IRootStore } from './rootStore';
 import { IGeoPackage } from './geoPackage';
@@ -26,24 +26,6 @@ export interface PackageInfo {
   tilesEst: number;
   minZoom?: number;
   maxZoom?: number;
-}
-
-const wmsOptions = EXPORTER_CONFIG.WMS_LAYER;
-const fullUrls : {[key : string] : string} = {
-  'WMS_LAYER': `${wmsOptions.URL}?${getParamsString(wmsOptions.PARAMS)}`,
-  'WMTS_LAYER': EXPORTER_CONFIG.WMTS_LAYER.URL,
-  'XYZ_LAYER': EXPORTER_CONFIG.XYZ_LAYER.URL,
-  'OSM_DEFAULT': 'NOT_VALID_URL'
-};
-
-function getParamsString(params: any) : string {
-  const paramArray = Object.keys(params).map((key) => `${key}=${params[key]}`);
-  return paramArray.join('&');
-}
-
-function getFullUrl() : string {
-  const activeLayer = EXPORTER_CONFIG.ACTIVE_LAYER;
-  return fullUrls[activeLayer];
 }
 
 export type ExporterResponse = ApiHttpResponse<ExportResult>;
@@ -78,7 +60,7 @@ export const exporterStore = types
         params.sizeEst = packInfo.sizeEst;
         params.tilesEst = packInfo.tilesEst;
         params.directoryName = 'test';
-        params.exportedLayers = [{exportType: 'raster', url: getFullUrl()}];
+        params.exportedLayers = [{exportType: 'raster', url: getLayerUrl()}];
         const coordinates = (snapshot.geojson as Polygon).coordinates[0];
         params.bbox = [coordinates[0][0], coordinates[0][1], coordinates[2][0], coordinates[2][1]];
